Updated version, now works.
Functionality has been implemented to interrogate the database for some meta data.
This version only performs the necessary operations to retrieve the password.
It is interesting to note that terminating the SQL query with '; --' worked but not reliably, while ' %23' worked very reliably. It may be interesting to investigate why this was happening. Would adding a space after the '; --' have helped? Can different 'frames' be automatically attempted?